packages/citadel-agents/pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "citadel-agents"
version = "0.1.0"
description = "Agent runtime framework — tool use, memory, planning, multi-agent orchestration"
authors = [{name = "Donald Havery", email = "dbhavery@gmail.com"}]
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = ["httpx>=0.28", "pyyaml>=6.0"]

[project.optional-dependencies]
anthropic = ["anthropic>=0.50"]
vector = ["citadel-vector>=0.1"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.26"]

[tool.setuptools.packages.find]
include = ["citadel_agents*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
